core: refactor run_command into commonwealth
